RFC 9022 - Adv Search Vorbereitungen
git-svn-id: http://10.205.116.236/svn/AIR@311 d2170373-ae6f-44b7-9150-2eec3990abfc

+++ branches/2.00/src/com/bayerbbs/applrepos/service/CiSearchParamsDTO.java
@@ -4,8 +4,8 @@ package com.bayerbbs.applrepos.service;
 public class CiSearchParamsDTO { //implements Serializable
 	private String cwid;
 	private String token;
-	private Long start;
-	private Long limit;
+	private Integer start;
+	private Integer limit;
 	private String sort;
 	private String dir;
 	
@@ -60,17 +60,17 @@ public class CiSearchParamsDTO { //implements Serializable
 		this.ciNameAliasQuery = ciNameAliasQuery;
 	}
 
-	public Long getStart() {
+	public Integer getStart() {
 		return start;
 	}
-	public void setStart(Long start) {
+	public void setStart(Integer start) {
 		this.start = start;
 	}
 
-	public Long getLimit() {
+	public Integer getLimit() {
 		return limit;
 	}
-	public void setLimit(Long limit) {
+	public void setLimit(Integer limit) {
 		this.limit = limit;
 	}
